+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [7.0.14] - 2025-04-17
+### Changed
+- Updated Setup Attendance panel [#4777](https://github.com/rokwire/illinois-app/issues/4777).
+- Reworked Event admin settings menu and panels [#5013](https://github.com/rokwire/illinois-app/issues/5013).
+
 ## [7.0.13] - 2025-04-16
 ### Changed
 - Improved accessibility for EventsHomePanel [#4417](https://github.com/rokwire/illinois-app/issues/4417).